% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_countries.R
\name{distinct_countries}
\alias{distinct_countries}
\title{Filter distinct report and partner countries}
\usage{
distinct_countries(data)
}
\arguments{
\item{data}{raw data}
}
\value{
a filtered tibble
}
\description{
Filter distinct report and partner countries
}
