% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{distinct_countries_for_dt}
\alias{distinct_countries_for_dt}
\title{Create output table filtering countries and year to be showed in DT}
\usage{
distinct_countries_for_dt(
  data_raw,
  input_origin,
  input_destination,
  input_year
)
}
\arguments{
\item{data_raw}{raw data}

\item{input_origin}{input origin}

\item{input_destination}{input destination}

\item{input_year}{input year}
}
\value{
a filtered tibble
}
\description{
Create output table filtering countries and year to be showed in DT
}
