% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_input_select_countries.R
\name{distinct_input_select_countries}
\alias{distinct_input_select_countries}
\title{Reorder country names and create a vector variable for the input panel}
\usage{
distinct_input_select_countries(data, countries_type)
}
\arguments{
\item{data}{raw data}

\item{countries_type}{reporter or partner countries}
}
\value{
an ordered vector
}
\description{
Reorder country names and create a vector variable for the input panel
}
