% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_input_select_years.R
\name{distinct_input_select_years}
\alias{distinct_input_select_years}
\title{Reorder years and create a vector variable for the input panel}
\usage{
distinct_input_select_years(data)
}
\arguments{
\item{data}{raw data}
}
\value{
an ordered vector
}
\description{
Reorder years and create a vector variable for the input panel
}
