% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_countries_by_input_select_countries}
\alias{filter_countries_by_input_select_countries}
\title{Title}
\usage{
filter_countries_by_input_select_countries(
  data_raw,
  input_origin,
  input_destination
)
}
\arguments{
\item{data_raw}{}

\item{input_origin}{}

\item{input_destination}{}
}
\value{

}
\description{
Title
}
