% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_countries_by_input_select_year.R
\name{filter_countries_by_input_select_year}
\alias{filter_countries_by_input_select_year}
\title{Filter values based on year}
\usage{
filter_countries_by_input_select_year(data_filtered_by_countries, input_year)
}
\arguments{
\item{data_filtered_by_countries}{a pre filtered tibble resulting from \code{filter_countries_by_input_select_countries()}}

\item{input_year}{Input year}
}
\value{
a refiltered tibble
}
\description{
Filter values based on year
}
