% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_year_by_input_select_year}
\alias{filter_year_by_input_select_year}
\title{Title}
\usage{
filter_year_by_input_select_year(data_raw, input_year)
}
\arguments{
\item{data_raw}{}

\item{input_year}{}
}
\value{

}
\description{
Title
}
