% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_by_input_colormap.R
\name{make_plot_by_input_colormap}
\alias{make_plot_by_input_colormap}
\title{High-order function to plot the final maps}
\usage{
make_plot_by_input_colormap(data_filtered, data_sf, input_colormap)
}
\arguments{
\item{data_filtered}{prefiltered tibble}

\item{data_sf}{sf data}

\item{input_colormap}{input colormap}
}
\value{
a map
}
\description{
High-order function to plot the final maps
}
