% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_vp_flow_by_input_year.R
\name{min_max_vp_flow_by_input_year}
\alias{min_max_vp_flow_by_input_year}
\title{Summarise minimum and maximum values of vp_flow for the choosed years
in the input panel}
\usage{
min_max_vp_flow_by_input_year(data, year)
}
\arguments{
\item{data}{filtered data}

\item{year}{choosed year in the input panel}
}
\value{
a tibble
}
\description{
Summarise minimum and maximum values of vp_flow for the choosed years
in the input panel
}
