% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{no_vp_flow}
\alias{no_vp_flow}
\title{Informs when countries doesn't have relationships}
\usage{
no_vp_flow(origin, destination)
}
\arguments{
\item{origin}{selected "Exporting countries"}

\item{destination}{select "Import countries"}
}
\value{
a plot
}
\description{
Informs when countries doesn't have relationships
}
