% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{plot_sf_map}
\alias{plot_sf_map}
\title{Create base map for the plot in the output panel}
\usage{
plot_sf_map(data_sf, filled_by)
}
\arguments{
\item{data_sf}{the sf object read by `read_sf_data()`}

\item{filled_by}{the colormap choosed in the input panel}
}
\value{
a sf object
}
\description{
Create base map for the plot in the output panel
}
