% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{summarise_vp_flow_all_years}
\alias{summarise_vp_flow_all_years}
\title{Summarise vp_flow data for all years}
\usage{
summarise_vp_flow_all_years(data)
}
\arguments{
\item{data}{filtered data by `distinct_countries()`}
}
\value{
a tibble
}
\description{
Used when all years are choosed in the input panel
}
