% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_countries_by_input_select_countries}
\alias{filter_countries_by_input_select_countries}
\title{Filter countries by input selection}
\usage{
filter_countries_by_input_select_countries(
  data_raw,
  input_origin,
  input_destination
)
}
\arguments{
\item{data_raw}{Data raw}

\item{input_origin}{Input reporter_countries}

\item{input_destination}{Input partner_countries}
}
\value{
a filtered tibble
}
\description{
Filter countries by input selection
}
