% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_countries_by_input_select_year}
\alias{filter_countries_by_input_select_year}
\title{Title}
\usage{
filter_countries_by_input_select_year(data_filtered_by_countries, input_year)
}
\arguments{
\item{data_filtered_by_countries}{}

\item{input_year}{}
}
\value{

}
\description{
Title
}
