% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_year_by_input_select_year}
\alias{filter_year_by_input_select_year}
\title{Filter and create a tibble with the minimum and maximum values of vp_flow 
to used at plot colors scale}
\usage{
filter_year_by_input_select_year(data_raw, input_year)
}
\arguments{
\item{data_raw}{data rae}

\item{input_year}{input year}
}
\value{
a filtered tibble
}
\description{
Filter and create a tibble with the minimum and maximum values of vp_flow 
to used at plot colors scale
}
