% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{make_plot_by_input_colormap}
\alias{make_plot_by_input_colormap}
\title{High-order function the plot the final maps}
\usage{
make_plot_by_input_colormap(
  data_filtered,
  data_sf,
  data_year,
  input_colormap,
  input_origin,
  input_destination
)
}
\arguments{
\item{data_filtered}{prefiltered tibble}

\item{data_sf}{sf data}

\item{data_year}{prefiltered tibble with min and max vp_flow values}

\item{input_colormap}{input colormap}

\item{input_origin}{input reporter_countries}

\item{input_destination}{input partner_countries}
}
\value{
a map
}
\description{
High-order function the plot the final maps
}
