% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{make_plot_by_input_colormap}
\alias{make_plot_by_input_colormap}
\title{Title}
\usage{
make_plot_by_input_colormap(
  data_filtered,
  data_sf,
  data_year,
  input_colormap,
  input_origin,
  input_destination
)
}
\arguments{
\item{data_filtered}{}

\item{data_sf}{}

\item{data_year}{}

\item{input_colormap}{}

\item{input_origin}{}

\item{input_destination}{}
}
\value{

}
\description{
Title
}
