% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{vp_flow_arrows_plot}
\alias{vp_flow_arrows_plot}
\title{Create the final plot}
\usage{
vp_flow_arrows_plot(
  virtual_pollinators_flow_filtered,
  base_world_map,
  vp_flow_year
)
}
\arguments{
\item{virtual_pollinators_flow_filtered}{the filtered data which already
passed by `distinct_countries()` and by `summarise_vp_flow_all_years()`,
if choosed in the input panel}

\item{base_world_map}{the sf object created by `plot_sf_map()`}

\item{vp_flow_year}{the filtered data created by `min_max_vp_flow_by_input_year()`}
}
\value{
a ggplot object: the final plot
}
\description{
Create the final plot
}
